Final Year Project/requirements.txt
Flask
numpy
pandas
scikit-learn
joblib
mysql-connector-python